Move config package one level up
