+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscan/orchestrator/config/IntegrationTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.controllers.config
+package uk.gov.hmcts.reform.bulkscan.orchestrator.config
 
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock
