+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscan/orchestrator/controllers/RootControllerTest.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.controllers.config.IntegrationTest;
+import uk.gov.hmcts.reform.bulkscan.orchestrator.config.IntegrationTest;
 
 import static io.restassured.RestAssured.given;
