+++ src/integrationTest/java/uk/gov/hmcts/reform/bulkscan/orchestrator/controllers/SupplementaryEvidenceCreatorTest.kt
@@ -13,9 +13,9 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.test.context.junit.jupiter.SpringExtension
-import uk.gov.hmcts.reform.bulkscan.orchestrator.controllers.config.Environment.caseEventUrl
-import uk.gov.hmcts.reform.bulkscan.orchestrator.controllers.config.Environment.getCaseUrl
-import uk.gov.hmcts.reform.bulkscan.orchestrator.controllers.config.IntegrationTest
+import uk.gov.hmcts.reform.bulkscan.orchestrator.config.Environment.caseEventUrl
+import uk.gov.hmcts.reform.bulkscan.orchestrator.config.Environment.getCaseUrl
+import uk.gov.hmcts.reform.bulkscan.orchestrator.config.IntegrationTest
 import java.io.File
 import java.util.concurrent.TimeUnit
